fastapi_x_vanillajs/pyproject.toml
[project]
name = "todo-by-mukund"
version = "0.1.0"
description = "FastAPI-based Todo application with user authentication and MongoDB integration."
authors = [
    { name = "Mukund Thorat", email = "mukundthorat.official@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.6",
    "bcrypt==4.3.0",
    "certifi>=2026.1.4",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.40.0",
]
